feat: add configuration for yazi
Added a new configuration entry for the Yazi terminal emulator in the
install.conf.yaml file to better organize and manage user
configurations. This enhancement simplifies the management of custom
settings for applications by consolidating them under the ~/.config
directory, making it easier to locate and modify configurations as
needed.