feat(nix): adds home manager to flake & hardcoded system
Update the `flake.lock` file to use the latest revisions and hashes
for `home-manager` and `nixpkgs` repositories. This ensures the
project uses the most recent versions of dependencies.

Modify `flake.nix` to include specific system architecture for each
host. This change allows for better handling of different system
architectures, ensuring configurations are correctly applied.

Add `ssh.nix` to `konoha` host configuration to enable and configure
OpenSSH service. This addition allows for remote access and management
of the `konoha` host.

Update `home-manager/home.nix` to include `pkgs` in the argument list,
ensuring the correct packages are used in home-manager configurations.

Signed-off-by: Marcelo Borges <git@marceloborges.dev>