feat(helix): add Helix editor configuration and integrate into setup
Introduce a configuration file for the Helix editor with a custom
theme and editor settings. Add Helix to the list of home-manager
packages to ensure it is installed. Update the install configuration
to create a symlink for the Helix configuration directory.

These changes enable the use of Helix as an editor with a predefined
setup, improving consistency and usability across environments.

Lets give a try to helix!