chore: update flake.lock dependencies
Update locked versions of `home-manager`, `nixpkgs`, and `rust-overlay`
to their latest revisions. This ensures the project uses the most
recent changes and improvements from upstream repositories. The updates
include changes to `lastModified`, `narHash`, and `rev` fields, which
reflect the new state of the dependencies. This helps maintain
compatibility and leverage bug fixes or new features provided by the
upstream sources.

Signed-off-by: Marcelo Borges <marcelo.borges@synopsys.com>